pyproject.toml
[tool.poetry]
name = "dev-backend-playground"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dev_backend_example"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
opentelemetry-instrumentation-fastapi = "^0.40b0"
opentelemetry-instrumentation-botocore = "^0.40b0"
opentelemetry-sdk-extension-aws = "^2.0.1"
opentelemetry-exporter-otlp = "^1.19.0"
opentelemetry-instrumentation-requests = "^0.40b0"
botocore = "^1.31.22"
boto3 = "^1.28.22"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.282"
black = "^23.7.0"
isort = "^5.12.0"
bandit = "^1.7.5"
pytest = "^7.4.0"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["ALL"]
ignore = ["I", "D", "ANN101", "A003"]
target-version = "py311"
line-length = 120
exclude = [".venv", "test_*.py"]
